Admin role added.

+++ app/Http/Controllers/GameController.php
@@ -58,6 +58,12 @@ class GameController extends Controller
 
     public function getUpdate(Request $request, $id)
     {
+        $user = User::findMe();
+
+        if (!$user || !$user->isAdmin()) {
+            return redirect('/');
+        }
+
         $game = Game::where(['id' => $id])->with([
             'gamesUsersA.user', 'gamesUsersB.user'
         ])->first();
@@ -86,6 +92,12 @@ class GameController extends Controller
 
     public function postUpdate(Request $request, $id)
     {
+        $user = User::findMe();
+
+        if (!$user || !$user->isAdmin()) {
+            return redirect('/');
+        }
+
         $this->validate($request, $this->validationRules());
         $game = Game::where('id', $id)->first();
 
@@ -100,6 +112,12 @@ class GameController extends Controller
 
     public function getDelete(Request $request, $id)
     {
+        $user = User::findMe();
+
+        if (!$user || !$user->isAdmin()) {
+            return redirect('/');
+        }
+
         $game = Game::findOrFail($id);
         $game->delete();
